requirements.txt
pyasic==0.48.2
influxdb_client[async]==1.39.0
python-decouple==3.8
